refactor(Error): 에러와 유효성 검사 클래스 분리 및 연결, 별도 폴더 생성
Error 폴더를 생성하였습니다.
- 에러와 유효성에 관련된 로직들을 이 폴더에 넣었습니다.
- exceptionHandler: 예외 처리를 하는 함수
- InputValidation: 입력값에 대한 검사를 진행하고 exceptionHandler를 호출하는 메서드를 모은 클래스
- BridgeError: Error를 상속받은 클래스로 에러 메시지를 출력한다.
OutPutView에 printError 메서드를 추가하였습니다.
