+++ docs/README.md
@@ -61,7 +61,7 @@
 - [x] 상수 분리하기
   - [x] 메시지
   - [x] 다리와 게임 관련 상수
-- [ ] 에러와 유효성 검사 클래스 분리 및 연결
+- [x] 에러와 유효성 검사 클래스 분리 및 연결
 - [x] 다리를 그리는 메서드 수정 및 분리
 
 ## 구조 및 전제 설정
