+++ src/Error/BridgeError.js
@@ -1,0 +1,10 @@
+const { printError } = require('../view/OutputView');
+
+class BridgeError extends Error {
+  constructor(message, ...error) {
+    super(...error);
+    printError(message);
+  }
+}
+
+module.exports = BridgeError;
