+++ src/view/OutputView.js
@@ -10,6 +10,10 @@ const OutputView = {
   printNewLine() {
     Console.print(OUTPUT_MESSAGE.new_line);
   },
+  printError(message) {
+    OutputView.printNewLine();
+    Console.print(message);
+  },
   /**
    * 현재까지 이동한 다리의 상태를 정해진 형식에 맞춰 출력한다.
    * <p>
